+++ src/main/java/menu/Application.java
@@ -1,6 +1,8 @@
 package menu;
 
 
+import menu.controller.MenuRecommendationController;
+
 public class Application {
     public static void main(String[] args) {
         MenuRecommendationController controller = new MenuRecommendationController();
