+++ src/main/java/menu/model/Menu.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.model;
 
 import java.util.Arrays;
 import java.util.List;
